packages/libraries/core/src/components/toggle/ods-toggle-controller.spec.ts
import { OdsClearLoggerSpy, OdsInitializeLoggerSpy, OdsLoggerSpyReferences } from '@ovhcloud/ods-testing/src';
import { Ods } from './../../configure/ods';
import { OdsThemeColorIntent } from '@ovhcloud/ods-theming';
import { OdsToggle } from './ods-toggle';
import { OdsToggleController } from './ods-toggle-controller';
import { OdsToggleMock } from './ods-toggle-mock';
import { OdsLogger } from '../../logger/ods-logger';

describe('spec:ods-toggle-controller', () => {
  let controller: OdsToggleController;
  let component: OdsToggle;
  let loggerSpyReferences: OdsLoggerSpyReferences;

  Ods.instance().logging(false);

  function setup(attributes: Partial<OdsToggle> = {}) {
    component = { ...component, ...attributes };
    controller = new OdsToggleController(component);
  }

  beforeEach(() => {
    component = new OdsToggleMock();

    const loggerMocked = new OdsLogger('myLoggerMocked');
    loggerSpyReferences = OdsInitializeLoggerSpy({
      loggerMocked: loggerMocked as never,
      spiedClass: OdsToggleController
    });
  });

  afterEach(() => {
    OdsClearLoggerSpy(loggerSpyReferences);
  });

  describe('methods', () => {
    beforeEach(() => {
      setup({
        color: OdsThemeColorIntent.default
      });
    });

    describe('methods:validateAttributes', () => {
      it('should not call console.warn', () => {
        controller.validateAttributes();
        expect(loggerSpyReferences.methodSpies.warn).toHaveBeenCalledTimes(0);
      });

      it('should call console.warn with wrong color', () => {
        const expected = 'The color attribute must have a value from [default, primary, text, accent, error, warning, success, info, promotion]';
        component.color = 'color' as OdsThemeColorIntent;

        controller.validateAttributes();
        expect(loggerSpyReferences.methodSpies.warn).toHaveBeenCalledWith(expected);
        expect(loggerSpyReferences.methodSpies.warn).toHaveBeenCalledTimes(1);
      });
    });
  });
});
